just some bug removed

+++ snmp4sdn/src/main/java/org/opendaylight/snmp4sdn/internal/Activator.java
@@ -64,7 +64,7 @@ public class Activator extends ComponentActivatorAbstractBase {
      */
     public void init() {
         Node.NodeIDType.registerIDType("SNMP", Long.class);
-        NodeConnector.NodeConnectorIDType.registerIDType("SNMP", Long.class, "SNMP");
+        NodeConnector.NodeConnectorIDType.registerIDType("SNMP", Short.class, "SNMP");
     }
 
     /**
